update not-found page
